# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Evaluation Service API
  version: 1.2.0
  description: |
    Full API contract for the current backend implementation (Phases 1-5).
servers:
  - url: http://localhost:8080
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/admin/notifications/rules:
    post:
      tags: [Notifications]
      summary: Create notification rule (admin)
      description: Enabled only when `features.enable-notification-rule-engine=true`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNotificationRuleRequest'
      responses:
        '201':
          description: Created rule
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationRuleResponse'
    get:
      tags: [Notifications]
      summary: List notification rules (admin)
      description: Enabled only when `features.enable-notification-rule-engine=true`.
      parameters:
        - in: query
          name: campaignId
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Rule list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NotificationRuleResponse'
  /api/v1/admin/notifications/rules/{id}:
    get:
      tags: [Notifications]
      summary: Get notification rule by id (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Rule
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationRuleResponse'
    put:
      tags: [Notifications]
      summary: Update notification rule (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateNotificationRuleRequest'
      responses:
        '200':
          description: Updated rule
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationRuleResponse'
    delete:
      tags: [Notifications]
      summary: Delete notification rule (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Deleted
  /api/v1/admin/notifications/templates:
    post:
      tags: [Notifications]
      summary: Create notification template (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNotificationTemplateRequest'
      responses:
        '201':
          description: Created template
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationTemplateResponse'
    get:
      tags: [Notifications]
      summary: List notification templates (admin)
      parameters:
        - in: query
          name: campaignId
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Template list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NotificationTemplateResponse'
  /api/v1/admin/notifications/templates/{id}:
    get:
      tags: [Notifications]
      summary: Get notification template by id (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Template
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationTemplateResponse'
    put:
      tags: [Notifications]
      summary: Update notification template (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateNotificationTemplateRequest'
      responses:
        '200':
          description: Updated template
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationTemplateResponse'
    delete:
      tags: [Notifications]
      summary: Delete notification template (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Deleted
  /api/v1/admin/notifications/templates/{id}/test-render:
    post:
      tags: [Notifications]
      summary: Render notification template body with variables (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TestNotificationTemplateRequest'
      responses:
        '200':
          description: Rendered template body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateRenderResponse'
...
  /api/v1/templates:
    post:
      tags: [Templates]
      summary: Create template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTemplateRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateResponse'
    get:
      tags: [Templates]
      summary: List templates
      parameters:
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 0
        - in: query
          name: size
          schema:
            type: integer
      responses:
        '200':
          description: Template list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TemplateResponse'
  /api/v1/templates/{id}:
    get:
      tags: [Templates]
      summary: Get template
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Template
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateResponse'
    put:
      tags: [Templates]
      summary: Update template
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTemplateRequest'
      responses:
        '200':
          description: Updated template
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateResponse'
    delete:
      tags: [Templates]
      summary: Delete template
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Deleted
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      in: path
      name: id
      required: true
      schema:
        type: string
  schemas:
    AnyObject:
      type: object
      additionalProperties: true
...
    CreateTemplateRequest:
      type: object
      required: [name]
      properties:
        name: { type: string }
        description: { type: string }
        category: { type: string }
        scoringMethod:
          type: string
          enum: [WEIGHTED_AVERAGE, SIMPLE_AVERAGE, MEDIAN, PERCENTILE_RANK, CUSTOM_FORMULA]
        sections:
          type: array
          items:
            $ref: '#/components/schemas/SectionRequest'
        customFormula: { type: string }
...
    UpdateTemplateRequest:
      type: object
      required: [name]
      properties:
        name: { type: string }
        description: { type: string }
        category: { type: string }
        scoringMethod:
          type: string
          enum: [WEIGHTED_AVERAGE, SIMPLE_AVERAGE, MEDIAN, PERCENTILE_RANK, CUSTOM_FORMULA]
        sections:
          type: array
          items:
            $ref: '#/components/schemas/SectionRequest'
        customFormula: { type: string }
    SectionRequest:
      type: object
      required: [title]
      properties:
        id: { type: string }
        title: { type: string }
        description: { type: string }
        orderIndex: { type: integer }
        weight: { type: number, format: double }
        questions:
          type: array
          items:
            $ref: '#/components/schemas/QuestionRequest'
    QuestionRequest:
      type: object
      required: [text, type]
      properties:
        id: { type: string }
        text: { type: string }
        type:
          type: string
          enum:
            - SINGLE_CHOICE
            - MULTIPLE_CHOICE
            - LIKERT_SCALE
            - OPEN_TEXT
            - NUMERIC_RATING
            - BOOLEAN
            - MATRIX
            - RANKING
            - NPS
            - FILE_UPLOAD
        orderIndex: { type: integer }
        required: { type: boolean }
        options:
          type: array
          items: { type: string }
        weight: { type: number, format: double }
        metadata:
          $ref: '#/components/schemas/AnyObject'
        conditionalLogic: { type: string }
    TemplateResponse:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string }
        category: { type: string }
        status:
          type: string
          enum: [DRAFT, PUBLISHED, DEPRECATED, ARCHIVED]
        currentVersion: { type: integer }
        scoringMethod:
          type: string
          enum: [WEIGHTED_AVERAGE, SIMPLE_AVERAGE, MEDIAN, PERCENTILE_RANK, CUSTOM_FORMULA]
        totalQuestions: { type: integer }
        sections:
          type: array
          items:
            $ref: '#/components/schemas/SectionResponse'
        createdBy: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    SectionResponse:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        description: { type: string }
        orderIndex: { type: integer }
        weight: { type: number, format: double }
        questions:
          type: array
          items:
            $ref: '#/components/schemas/QuestionResponse'
    QuestionResponse:
      type: object
      properties:
        id: { type: string }
        text: { type: string }
        type: { type: string }
        orderIndex: { type: integer }
        required: { type: boolean }
        options:
          type: array
          items: { type: string }
        weight: { type: number, format: double }
        metadata:
          $ref: '#/components/schemas/AnyObject'
        conditionalLogic: { type: string }
...
    CreateNotificationRuleRequest:
      type: object
      required: [campaignId, ruleCode, triggerType, audience, channel]
      properties:
        campaignId: { type: string }
        ruleCode: { type: string }
        triggerType:
          type: string
          enum: [EVALUATION_SUBMITTED, CAMPAIGN_CLOSED, SCHEDULED]
        audience:
          type: string
          enum: [EVALUATOR, EVALUATEE, CAMPAIGN_OWNER, CUSTOM]
        channel:
          type: string
          enum: [IN_APP, EMAIL, WEBHOOK, SMS]
        scheduleExpr: { type: string, nullable: true }
        enabled: { type: boolean }
        config:
          $ref: '#/components/schemas/AnyObject'
    UpdateNotificationRuleRequest:
      type: object
      properties:
        triggerType:
          type: string
          enum: [EVALUATION_SUBMITTED, CAMPAIGN_CLOSED, SCHEDULED]
          nullable: true
        audience:
          type: string
          enum: [EVALUATOR, EVALUATEE, CAMPAIGN_OWNER, CUSTOM]
          nullable: true
        channel:
          type: string
          enum: [IN_APP, EMAIL, WEBHOOK, SMS]
          nullable: true
        scheduleExpr: { type: string, nullable: true }
        enabled: { type: boolean, nullable: true }
        config:
          $ref: '#/components/schemas/AnyObject'
    NotificationRuleResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        campaignId: { type: string }
        ruleCode: { type: string }
        triggerType: { type: string }
        audience: { type: string }
        channel: { type: string }
        scheduleExpr: { type: string, nullable: true }
        enabled: { type: boolean }
        config:
          $ref: '#/components/schemas/AnyObject'
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CreateNotificationTemplateRequest:
      type: object
      required: [templateCode, name, channel, body]
      properties:
        campaignId: { type: string, nullable: true }
        templateCode: { type: string }
        name: { type: string }
        channel:
          type: string
          enum: [IN_APP, EMAIL, WEBHOOK, SMS]
        subject: { type: string, nullable: true }
        body: { type: string }
        requiredVariables:
          type: array
          items: { type: string }
        status:
          type: string
          enum: [DRAFT, PUBLISHED, ARCHIVED]
          nullable: true
    UpdateNotificationTemplateRequest:
      type: object
      properties:
        name: { type: string, nullable: true }
        channel:
          type: string
          enum: [IN_APP, EMAIL, WEBHOOK, SMS]
          nullable: true
        subject: { type: string, nullable: true }
        body: { type: string, nullable: true }
        requiredVariables:
          type: array
          items: { type: string }
          nullable: true
        status:
          type: string
          enum: [DRAFT, PUBLISHED, ARCHIVED]
          nullable: true
    TestNotificationTemplateRequest:
      type: object
      required: [recipient]
      properties:
        recipient: { type: string }
        variables:
          $ref: '#/components/schemas/AnyObject'
    NotificationTemplateResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        campaignId: { type: string, nullable: true }
        templateCode: { type: string }
        name: { type: string }
        channel: { type: string }
        subject: { type: string, nullable: true }
        body: { type: string }
        requiredVariables:
          type: array
          items: { type: string }
        status: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    TemplateRenderResponse:
      type: object
      properties:
        recipient: { type: string }
        renderedBody: { type: string }